+++ ZenUnitTests/TestRunners/PreamblePrinterTests.cpp
@@ -23,7 +23,7 @@ namespace ZenUnit
          _watch.reset(watchMock = new WatchMock);
       }
       ZENMOCK_NONVOID1_CONST(string, MakeThirdLinePrefix, size_t)
-      ZENMOCK_NONVOID2_CONST(string, MakeThirdLineSuffix, bool, unsigned)
+      ZENMOCK_NONVOID2_CONST(string, MakeThirdLineSuffix, bool, unsigned short)
    } _preamblePrinterSelfMocked;
 
    PreamblePrinter _preamblePrinter;
@@ -55,7 +55,7 @@ namespace ZenUnit
       ZenUnitArgs zenUnitArgs;
       zenUnitArgs.commandLine = Random<string>();
       zenUnitArgs.random = Random<bool>();
-      zenUnitArgs.randomseed = Random<unsigned>();
+      zenUnitArgs.randomseed = Random<unsigned short>();
       //
       _preamblePrinterSelfMocked.PrintOpeningThreeLines(zenUnitArgs, &testClassRunnerRunnerMock);
       //
@@ -90,10 +90,10 @@ namespace ZenUnit
    }
 
    TEST3X3(MakeThirdLineSuffix_ReturnsRandomSeedIfRandomModeOtherwiseEmptyString,
-      const string& expectedReturnValue, bool random, unsigned randomseed,
-      "", false, NA<unsigned>(),
-      " (random seed 0)", true, 0u,
-      " (random seed 1)", true, 1u)
+      const string& expectedReturnValue, bool random, unsigned short randomseed,
+      "", false, NA<unsigned short>(),
+      " (random seed 0)", true, static_cast<unsigned short>(0),
+      " (random seed 1)", true, static_cast<unsigned short>(1))
    {
       const string thirdLineSuffix = _preamblePrinter.MakeThirdLineSuffix(random, randomseed);
       ARE_EQUAL(expectedReturnValue, thirdLineSuffix);
